backend/requirements.txt
Django==4.1.10
django-cors-headers==4.3.0
psycopg2==2.9.9
Pillow==10.1.0
djangorestframework==3.14.0
PyJWT==2.8.0